[Refactor] 채팅 관련 기능 수정 - 채팅컴포넌트 프로필 이미지 로드 안되는 문제 해결 - 채팅방 입장 시 빈 알람 뜨는 문제 해결 - 채팅을 길게 쳤을때 의도치 않게 줄바꿈 되는 문제 해결 - 채팅방 입장 시 최신채팅내용 10개가 밑에서부터 위로 불러와지게 해결 - 같은 채팅방 여러번 입장 시 중복 구독을 하게 돼서 중복된 구독수 만큼 채팅컴포넌트가 나왔던 문제 해결
