+++ frontend/src/stores/useStompStore.js
@@ -4,7 +4,7 @@ import { useMessageStore } from "@/stores/useMessageStore";
 import { toast } from 'vue3-toastify';
 import 'vue3-toastify/dist/index.css';
 
-// const storedToken = localStorage.getItem("accessToken");
+const storedToken = localStorage.getItem("accessToken");
 const timeout = 10000;
 const backend = process.env.VUE_APP_WS_ENDPOINT;
 let retry = 0;
@@ -12,34 +12,49 @@ let retry = 0;
 
 export const useStompStore = defineStore("stomp", {
     state: () => ({
-        chatStomp: {}
+        chatStomp: null,
+        subscription: null
     }),
     actions: {
-        roomConnect(chatRoomId, router) {
-            window.localStorage.setItem("chatRoomId", chatRoomId);
-            const stomp = Stomp.client(`${backend}/chat`);
-            stomp.connect({}, frame => {
-                stomp.connected = true;
-                console.log('소켓 연결 성공', frame);
-                toast('채팅방에 접속하였습니다.', {
-                    timeout: timeout,
+        async roomConnect(chatRoomId, router) {
+            if(this.subscription !== null){
+                this.subscription.unsubscribe();
+            }
+                
+                
+             
+                window.localStorage.setItem("chatRoomId", chatRoomId);
+                const stomp = Stomp.client(`${backend}/chat`);
+                stomp.connect({}, frame => {
+                    stomp.connected = true;
+                    console.log('소켓 연결 성공', frame);
+                    toast('채팅방에 접속하였습니다.', {
+                        timeout: timeout,
+                    })
+                    this.chatStomp = stomp;
+                    this.subscription = stomp.subscribe("/sub/room/" + chatRoomId, res => {
+                        console.log("연결 후 채팅방 아이디", chatRoomId);
+                        console.log(res);
+                        console.log("구독으로 받은 메시지입니다.", res.body);
+                        useMessageStore().addMessage(JSON.parse(res.body));
+                    })
+                    
+                }, error => {
+                    console.log(error);
+                    console.log('=======에러발생=======');
+                    stomp.connected = false;
+                    console.log(error.code);
+                    if (error.code === 1001 || error.code === 1002 || error.code === 1006) {
+                        this.retrySocketConnect(error, chatRoomId, router);
+                    }
                 })
-                this.chatStomp = stomp;
-                stomp.subscribe("/sub/room/" + chatRoomId, res => {
-                    console.log("연결 후 채팅방 아이디", chatRoomId);
-                    console.log(res);
-                    console.log("구독으로 받은 메시지입니다.", res.body);
-                    useMessageStore().addMessage(JSON.parse(res.body));
-                })
-            }, error => {
-                console.log(error);
-                console.log('=======에러발생=======');
-                stomp.connected = false;
-                console.log(error.code);
-                if (error.code === 1001 || error.code === 1002 || error.code === 1006) {
-                    this.retrySocketConnect(error, chatRoomId, router);
-                }
-            })
+                 
+                
+
+                
+                useMessageStore().recvList = []
+                await useMessageStore().getChatList(chatRoomId, storedToken, 1, 10); 
+                   
         },
         retrySocketConnect(error, chatRoomId, router) {
             toast.error('채팅방과 연결이 끊어졌습니다.', {
@@ -56,9 +71,4 @@ export const useStompStore = defineStore("stomp", {
             }
         }
     },
-    getters: {
-        getAllMessage(state) {
-            return state.recvList;
-        }
-    }
 })
